database/data-generation/requirements.txt
kafka-python==2.0.2
python-dotenv